refactor search function into a helper and combine partial with full searches
