+++ app-web/src/pages/index.js
@@ -4,7 +4,8 @@ import queryString from 'query-string';
 import shortid from 'shortid';
 import { connect } from 'react-redux';
 import { REACT_SCROLL } from '../constants/ui';
-import { flattenGatsbyGraphQL } from '../utils//dataHelpers';
+import { flattenGatsbyGraphQL } from '../utils/dataHelpers';
+import { getSearchResults } from '../utils/helpers';
 import * as actions from '../store/actions';
 
 import styles from './index.module.css';
@@ -15,7 +16,6 @@ import Loading from '../components/UI/Loading/Loading';
 import Layout from '../hoc/Layout';
 import Cards from '../components/Cards/Cards';
 import Masthead from '../components/Home/Masthead';
-import CollectionCard from '../components/Cards/Card/Collection';
 
 // selectors from reselect
 import {
@@ -70,29 +70,7 @@ export class Index extends PureComponent {
    */
   async getSearchResults(query) {
     const lunr = await window.__LUNR__.__loaded;
-    const lunrIndex = lunr.en;
-    let results = [];
-    let searchQuery = `*${query}*`;
-    // attempt to search by parsing query into fields
-    try {
-      results = lunrIndex.index.search(searchQuery);
-    } catch (e) {
-      // if that fails treat query as plain text and attempt search again
-      results = lunrIndex.index.query(function() {
-        this.term(searchQuery);
-      });
-    }
-    // search results is an array of reference keys
-    // we need to map those to the index store to get the actual
-    // node ids
-    const searchResultsMap = results
-      .map(({ ref }) => lunrIndex.store[ref])
-      .reduce((obj, result) => {
-        obj[result.id] = { ...result };
-        return obj;
-      }, {});
-
-    return searchResultsMap;
+    return getSearchResults(query, lunr);
   }
 
   render() {
