+++ app-web/src/utils/helpers.js
@@ -58,3 +58,41 @@ export const mapPagePathToResourceTypeConst = pathname => {
   // conver to upper case so we can access resource type enum properties
   return RESOURCE_TYPES[trimmedPath.toUpperCase()];
 };
+
+/**
+ * gets search results from lunr
+ * @param {String} query the search string
+ */
+export const getSearchResults = async (query, lunr) => {
+  const lunrIndex = lunr.en;
+  let results = [];
+  // search results by a partial query using wild cards
+  let partialResults = [];
+  let searchQueryPartial = `*${query}*`;
+  // attempt to search by parsing query into fields
+  try {
+    partialResults = lunrIndex.index.search(searchQueryPartial);
+    results = lunrIndex.index.search(query);
+  } catch (e) {
+    // if that fails treat query as plain text and attempt search again
+    partialResults = lunrIndex.index.query(function() {
+      this.term(searchQueryPartial);
+    });
+    results = lunrIndex.index.query(function() {
+      this.term(searchQueryPartial);
+    });
+  }
+  // combine all partial search results with full search results
+  results = results.concat(partialResults);
+  // search results is an array of reference keys
+  // we need to map those to the index store to get the actual
+  // node ids
+  const searchResultsMap = results
+    .map(({ ref }) => lunrIndex.store[ref])
+    .reduce((obj, result) => {
+      obj[result.id] = { ...result };
+      return obj;
+    }, {});
+
+  return searchResultsMap;
+};
